+++ postgres-types/Cargo.toml
@@ -49,4 +49,4 @@ uuid-1 = { version = "1.0", package = "uuid", optional = true }
 time-02 = { version = "0.2", package = "time", optional = true }
 time-03 = { version = "0.3", package = "time", default-features = false, optional = true }
 
-smol_str = { version = "0.1.23", default-features = false, optional = true }
+smol_str-01 = { version = "0.1.23", package = "smol_str", default-features = false, optional = true }
